chore(node): Move `NpmConfig` to a dedicated file
Prepare for enhancing `NpmConfig`.
